renamed data method to get-claim-summary
